chore(team): split cancel run manager tests
